+++ SyncClipboard/Env.cs
@@ -6,6 +6,7 @@ namespace SyncClipboard
     {
         public const string VERSION = "1.3.8";
         internal static readonly string Directory = System.Windows.Forms.Application.StartupPath;
+        internal static readonly string LOCAL_FILE_FOLDER = FullPath("file");
         internal static string FullPath(string relativePath)
         {
             return Path.Combine(Directory, relativePath);