+++ SyncClipboard/Service/SyncService/SyncService.cs
@@ -6,7 +6,7 @@ namespace SyncClipboard.Service
     {
         internal const string REMOTE_RECORD_FILE = "SyncClipboard.json";
         internal const string REMOTE_FILE_FOLDER = "file";
-        internal static readonly string LOCAL_FILE_FOLDER = Env.FullPath("file");
+        internal static readonly string LOCAL_FILE_FOLDER = Env.LOCAL_FILE_FOLDER;
         internal static Mutex remoteProfilemutex = new Mutex();
         internal static Mutex localProfilemutex = new Mutex();
         public const string PULL_START_ENENT_NAME = "PULL_START_ENENT";